(fix): Address code review findings — SSRF bypass, send checks, Console logging, SMTP constants, TLSContext rename
- Block IPv6-mapped IPv4 addresses (::ffff:) in SSRF validation
- Check send() return values in TCP fast path, forward goroutine, and SMTP forwarding
- Fix goroutine resource leak in TCP SwooleCoroutine error path
- Replace all echo/error_log with Utopia\Console (log, success, info, error)
- Extract SMTP magic numbers into class constants (RECV_BUFFER, PACKAGE_MAX_LENGTH, etc.)
- Rename TlsContext → TLSContext (acronym casing convention)
- Rename getTlsContext → getTLSContext in Config and all callers
- Add workerStart callback support to HTTP SwooleCoroutine
- Normalize timeout types to float across all Config classes
- Add utopia-php/console dependency, remove unused utopia-php/cli

+++ src/Server/HTTP/Config.php
@@ -26,7 +26,7 @@ class Config
         public readonly bool $parseFiles = false,
         public readonly bool $compression = false,
         public readonly int $logLevel = SWOOLE_LOG_ERROR,
-        public readonly int $timeout = 30,
+        public readonly float $timeout = 30.0,
         public readonly float $connectTimeout = 5.0,
         public readonly bool $keepAlive = true,
         public readonly int $poolSize = 1024,
